Finished first draft and added tests file

+++ src/Circle.php
@@ -7,16 +7,25 @@
 	class Circle {
 
 		private $api_key;
+		private $community_id;
 
 		public function __construct(string $api_key) {
 
-			$this->api_key = $token;
+			$this->api_key = $api_key;
 			$this->curl = new Curl();
-			$curl->setHeader('Authorization', "Token {$this->api_key}");
+			$this->curl->setHeader('Authorization', "Token {$this->api_key}");
 			$this->baseUrl = 'https://app.circle.so/api/v1/';
 		}
 
-		function toAscii($str, $replace = [], $delimiter = '-') {
+		public function setCommunity($id) {
+			$this->community_id = $id;
+		}
+
+		public function getCommunity() {
+			return $this->community_id;
+		}
+
+		public function toAscii($str, $replace = [], $delimiter = '-') {
 			setlocale(LC_ALL, 'en_US.UTF8');
 			# Remove spaces
 			if( !empty($replace) ) {
@@ -62,7 +71,10 @@
 
 		// Space Groups
 
-		public function spaceGroups($community_id) {
+		public function spaceGroups($community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
 
 			$this->curl->get($this->baseUrl . "/space_groups?community_id={$community_id}");
 
@@ -70,9 +82,12 @@
 			return $response;
 		}
 
-		public function spaceGroup($space_group_id, $community_id) {
+		public function spaceGroup($space_group_id, $community_id = 0) {
 
-			$this->curl->post($this->baseUrl . "/space_groups/{$space_group_id}?community_id={$community_id}");
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
+
+			$this->curl->get($this->baseUrl . "/space_groups/{$space_group_id}?community_id={$community_id}");
 
 			$response = json_decode($this->curl->response);
 			return $response;
@@ -80,7 +95,10 @@
 
 		// Space Group Members
 
-		public function addSpaceGroupMember($email, $space_group_id, $community_id) {
+		public function addSpaceGroupMember($email, $space_group_id, $community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
 
 			$this->curl->post($this->baseUrl . "/space_group_members?email={$email}&space_group_id={$space_group_id}&community_id={$community_id}");
 
@@ -88,7 +106,10 @@
 			return $response;
 		}
 
-		public function removeSpaceGroupMember($email, $space_group_id, $community_id) {
+		public function removeSpaceGroupMember($email, $space_group_id, $community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
 
 			$this->curl->delete($this->baseUrl . "/space_group_members?email={$email}&space_group_id={$space_group_id}&community_id={$community_id}");
 
@@ -96,7 +117,10 @@
 			return $response;
 		}
 
-		public function spaceGroupMember($email, $space_group_id, $community_id) {
+		public function spaceGroupMember($email, $space_group_id, $community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
 
 			$this->curl->get($this->baseUrl . "/space_group_member?community_id={$community_id}&email={$email}&space_group_id={$space_group_id}");
 
@@ -106,15 +130,29 @@
 
 		// Spaces
 
-		public function spaces($community_id, $sort = 'active', $per_page = 60, $page = 1) {
+		public function spaces($options = [], $community_id = 0) {
 
-			$this->curl->get($this->baseUrl . "spaces?community_id={$community_id}&sort={$sort}&per_page={$per_page}&page={$page}");
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
+
+			$params = [
+				'sort' => 'active',
+				'per_page' => 60,
+				'page' => 1
+			];
+
+			$params = array_merge($params, $options);
+
+			$this->curl->get($this->baseUrl . "spaces?community_id={$community_id}&sort={$sort}&per_page={$per_page}&page={$page}" . http_build_query($params));
 
 			$response = json_decode($this->curl->response);
 			return $response;
 		}
 
-		public function space($space_id, $community_id) {
+		public function space($space_id, $community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
 
 			$this->curl->get($this->baseUrl . "spaces/{$space_id}?community_id={$community_id}");
 
@@ -122,7 +160,10 @@
 			return $response;
 		}
 
-		public function createSpace($community_id, $space_group_id, $name, $options = []) {
+		public function createSpace($space_group_id, $name, $options = [], $community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
 
 			$params = [
 				'is_private' => false,
@@ -149,7 +190,10 @@
 
 		// Space Members
 
-		public function addSpaceMember($email, $space_id, $community_id) {
+		public function addSpaceMember($email, $space_id, $community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
 
 			$this->curl->post($this->baseUrl . "space_members?email={$email}&space_id={$space_id}&community_id={$community_id}");
 
@@ -157,7 +201,10 @@
 			return $response;
 		}
 
-		public function removeSpaceMember($email, $space_id, $community_id) {
+		public function removeSpaceMember($email, $space_id, $community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
 
 			$this->curl->delete($this->baseUrl . "space_members?email={$email}&space_id={$space_id}&community_id={$community_id}");
 
@@ -167,7 +214,10 @@
 
 		// Posts
 
-		public function posts($space_id, $community_id, $options = []) {
+		public function posts($space_id, $options = [], $community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
 
 			$params = [
 				'sort' => 'latest',
@@ -183,7 +233,10 @@
 			return $response;
 		}
 
-		public function post($post_id, $community_id) {
+		public function post($post_id, $community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
 
 			$this->curl->get($this->baseUrl . "posts/{$post_id}?community_id={$community_id}");
 
@@ -191,10 +244,12 @@
 			return $response;
 		}
 
-		public function createPost($space_id, $community_id, $name, $body, $options = []) {
+		public function createPost($space_id, $name, $body, $options = [], $community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
 
 			$params = [];
-
 			$params = array_merge($params, $options);
 
 			$this->curl->post($this->baseUrl . "posts/?community_id={$community_id}&space_id={$space_id}" . http_build_query($params));
@@ -203,7 +258,10 @@
 			return $response;
 		}
 
-		public function destroyPost($post_id, $space_id, $community_id) {
+		public function destroyPost($post_id, $space_id, $community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
 
 			$this->curl->delete($this->baseUrl . "posts/{$post_id}?community_id={$community_id}&space_id={$space_id}");
 
@@ -213,7 +271,7 @@
 
 		//Comments
 
-		public function comments($space_id, $community_id, $post_id = false) {
+		public function comments($space_id, $post_id = false, $community_id = 0) {
 
 			$this->curl->get($this->baseUrl . "comments?community_id={$community_id}&space_id={$space_id}" . ($post_id ? "&post_id={$post_id}" : ''));
 
@@ -221,7 +279,10 @@
 			return $response;
 		}
 
-		public function comment($comment_id, $community_id) {
+		public function comment($comment_id, $community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
 
 			$this->curl->get($this->baseUrl . "comments/{$comment_id}?community_id={$community_id}");
 
@@ -229,7 +290,10 @@
 			return $response;
 		}
 
-		public function createComment($space_id, $community_id, $body, $post_id = false, $options = []) {
+		public function createComment($space_id, $body, $options = [], $community_id = 0, $post_id = false) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
 
 			$params = [];
 			if($post_id) $params['post_id'] = $post_id;
@@ -242,9 +306,281 @@
 			return $response;
 		}
 
-		public function destroyComment($comment_id, $community_id) {
+		public function destroyComment($comment_id, $community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
 
 			$this->curl->delete($this->baseUrl . "comments/{$comment_id}?community_id={$community_id}");
+
+			$response = json_decode($this->curl->response);
+			return $response;
+		}
+
+		// Members
+
+		public function members($options = []) {
+
+			$params = [
+				'sort' => 'latest',
+				'per_page' => 10,
+				'page' => 1,
+				'status' => 'active'
+			];
+
+			$params = array_merge($params, $options);
+
+			$this->curl->get($this->baseUrl . "community_members?sort=latest&per_page=2&page=1&" . http_build_query($params));;
+
+			$response = json_decode($this->curl->response);
+			return $response;
+		}
+
+		public function member($member_id, $community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
+
+			$this->curl->get($this->baseUrl . "community_members/{$member_id}?community_id={$community_id}");
+
+			$response = json_decode($this->curl->response);
+			return $response;
+		}
+
+		public function inviteMember($email, $name, $options = [], $community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
+
+			$params = [];
+			$params = array_merge($params, $options);
+
+			$this->curl->post($this->baseUrl . "community_members?email={$email}&name={$name}&community_id={$community_id}&" . http_build_query($params));
+
+			$response = json_decode($this->curl->response);
+			return $response;
+		}
+
+		public function removeMember($email, $community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
+
+			$this->curl->delete($this->baseUrl . "community_members?community_id={$community_id}&email={$email}");
+
+			$response = json_decode($this->curl->response);
+			return $response;
+		}
+
+		public function updateMember($member_id, $options = [], $community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
+
+			$params = [];
+			$params = array_merge($params, $options);
+
+			$this->curl->put($this->baseUrl . "community_members/{$member_id}?community_id={$community_id}&" . http_build_query($params));
+
+			$response = json_decode($this->curl->response);
+			return $response;
+		}
+
+		public function searchMember($email, $community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
+
+			$this->curl->get($this->baseUrl . "community_members/search?community_id={$community_id}&email={$email}");
+
+			$response = json_decode($this->curl->response);
+			return $response;
+		}
+
+		// Member Tag
+
+		public function memberTags($community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
+
+			$this->curl->get($this->baseUrl . "member_tags?community_id={$community_id}");
+
+			$response = json_decode($this->curl->response);
+			return $response;
+		}
+
+		public function memberTag($member_tag_id, $community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
+
+			$this->curl->get($this->baseUrl . "member_tags/{$member_tag_id}?community_id={$community_id}");
+
+			$response = json_decode($this->curl->response);
+			return $response;
+		}
+
+		// Tagged Members
+
+		public function taggedMembers() {
+
+			$this->curl->get($this->baseUrl . "tagged_members");
+
+			$response = json_decode($this->curl->response);
+			return $response;
+		}
+
+		public function taggedMember($tagged_member_id) {
+
+			$this->curl->get($this->baseUrl . "tagged_members/{$tagged_member_id}");
+
+			$response = json_decode($this->curl->response);
+			return $response;
+		}
+
+		public function tagMember($email, $member_tag_id) {
+
+			$this->curl->post($this->baseUrl . "tagged_members?user_email={$email}&member_tag_id={$member_tag_id}");
+
+			$response = json_decode($this->curl->response);
+			return $response;
+		}
+
+		public function untagMember($email, $member_tag_id, $community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
+
+			$this->curl->delete($this->baseUrl . "tagged_members?user_email={$email}&member_tag_id={$member_tag_id}&community_id={$community_id}");
+
+			$response = json_decode($this->curl->response);
+			return $response;
+		}
+
+		// Group Messages
+
+		public function startGroupChat($message_body, $community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
+
+			$this->curl->post($this->baseUrl . "messages?community_id={$community_id}&message_body={$message_body}");
+
+			$response = json_decode($this->curl->response);
+			return $response;
+		}
+
+		// Direct Messages
+
+		public function sendMessage($email, $message_body, $community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
+
+			$this->curl->post($this->baseUrl . "messages?community_id={{community_id}}&user_email={$email}&message_body={$message_body}");
+
+			$response = json_decode($this->curl->response);
+			return $response;
+		}
+
+		// Member Subscriptions
+
+		public function memberSubscriptions($community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
+
+			$this->curl->get($this->baseUrl . "community_member_subscriptions?community_id={$community_id}");
+
+			$response = json_decode($this->curl->response);
+			return $response;
+		}
+
+		// Events
+
+		public function events() {
+
+			$this->curl->get($this->baseUrl . "events");
+
+			$response = json_decode($this->curl->response);
+			return $response;
+		}
+
+		public function createEvent($event, $space_id, $community_id = 0) {
+
+			$this->curl->post($this->baseUrl . "events?community_id={$community_id}&space_id={$space_id}", $event);
+
+			$response = json_decode($this->curl->response);
+			return $response;
+		}
+
+		public function updateEvent($event, $space_id, $community_id = 0) {
+
+			$this->curl->put($this->baseUrl . "events?community_id={$community_id}&space_id={$space_id}", $event);
+
+			$response = json_decode($this->curl->response);
+			return $response;
+		}
+
+		// Event Attendees
+
+		public function eventAttendees($event_id, $options = [], $community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
+
+			$params = [
+				'community_id' => $community_id,
+				'event_id' => $event_id
+			];
+
+			$params = array_merge($params, $options);
+
+			$this->curl->get($this->baseUrl . "event_attendees");
+
+			$response = json_decode($this->curl->response);
+			return $response;
+		}
+
+		public function rsvpMember($email, $event_id, $community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
+
+			$this->curl->post($this->baseUrl . "event_attendees", [
+				'community_id' => $community_id,
+				'event_id' => $event_id,
+				'email' => $email
+			]);
+
+			$response = json_decode($this->curl->response);
+			return $response;
+		}
+
+		public function unrsvpMember($email, $event_id, $community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
+
+			$this->curl->delete($this->baseUrl . "event_attendees", [
+				'community_id' => $community_id,
+				'event_id' => $event_id,
+				'email' => $email
+			]);
+
+			$response = json_decode($this->curl->response);
+			return $response;
+		}
+
+		// Member Charges
+
+		public function memberCharges($community_id = 0) {
+
+			$community_id = $community_id ?: $this->community_id;
+			if(!$community_id) return false;
+
+			$this->curl->get($this->baseUrl . "community_member_charges?community_id={$community_id}");
 
 			$response = json_decode($this->curl->response);
 			return $response;
